+++ changelog.md
@@ -35,7 +35,7 @@ All notable changes to this project will be documented in this file.
 - Passing `None` to comparison functions like `greater()`, `less()`, etc. used to raise an exception, which caused a test failure and stopped the test. 
  - That's unexpected with a check function. 
  - So now the check fails with a meaningful error message and continues with the rest of the test.
-
+- (typing) Allow numpy floats with comparison functions to pass mypy.
 
 ## [2.7.4] - 2026-Mar-03